plugin/commands/nas/nas.go
package nas

import (
	"github.com/IBM-Cloud/ibm-cloud-cli-sdk/plugin"
	"github.com/spf13/cobra"

	. "github.ibm.com/SoftLayer/softlayer-cli/plugin/i18n"
	"github.ibm.com/SoftLayer/softlayer-cli/plugin/metadata"
)

func SetupCobraCommands(sl *metadata.SoftlayerCommand) *cobra.Command {
	cobraCmd := &cobra.Command{
		Use:   "nas",
		Short: T("Classic infrastructure Network Attached Storage"),
		RunE:  nil,
	}
	cobraCmd.AddCommand(NewListCommand(sl).Command)
	cobraCmd.AddCommand(NewCredentialsCommand(sl).Command)
	return cobraCmd
}

func NasNetworkStorageNamespace() plugin.Namespace {
	return plugin.Namespace{
		ParentName:  "sl",
		Name:        "nas",
		Description: T("Classic infrastructure Network Attached Storage"),
	}
}
